rename type to group
