+++ example/src/main/java/ru/mintrocket/mintpermissions/dialogs_flow/DialogsViewModel.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import ru.mintrocket.lib.mintpermissions.MintPermissionsController
-import ru.mintrocket.lib.mintpermissions.flows.internal.FlowResultStatus
+import ru.mintrocket.lib.mintpermissions.flows.models.FlowResultStatus
 import ru.mintrocket.lib.mintpermissions.flows.internal.MintPermissionsFlowZygote
 import ru.mintrocket.mintpermissions.common.Event
