+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/MintPermissionsFlow.kt
@@ -3,6 +3,7 @@ package ru.mintrocket.lib.mintpermissions.flows
 import android.app.Application
 import ru.mintrocket.lib.mintpermissions.MintPermissionsConfig
 import ru.mintrocket.lib.mintpermissions.flows.internal.MintPermissionsFlowZygote
+import ru.mintrocket.lib.mintpermissions.flows.models.FlowConfig
 import ru.mintrocket.lib.mintpermissions.models.MintPermission
 
 object MintPermissionsFlow {
@@ -11,7 +12,10 @@ object MintPermissionsFlow {
         MintPermissionsFlowZygote.dialogsFlow
     }
 
-    fun createPlainFlow(permissions: List<MintPermission>): MintPermissionsPlainFlow {
+    fun createPlainFlow(
+        permissions: List<MintPermission>,
+        config: FlowConfig? = null
+    ): MintPermissionsPlainFlow {
         return MintPermissionsFlowZygote.createPlainFlow(permissions)
     }
