+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/MintPermissionsPlainFlow.kt
@@ -1,7 +1,7 @@
 package ru.mintrocket.lib.mintpermissions.flows
 
 import kotlinx.coroutines.flow.Flow
-import ru.mintrocket.lib.mintpermissions.flows.internal.FlowResultStatus
+import ru.mintrocket.lib.mintpermissions.flows.models.FlowResultStatus
 import ru.mintrocket.lib.mintpermissions.models.MintPermissionStatus
 
 interface MintPermissionsPlainFlow {
