+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/models/DialogRequest.kt
@@ -5,14 +5,8 @@ import ru.mintrocket.lib.mintpermissions.flows.ui.DialogContentMapper
 import ru.mintrocket.lib.mintpermissions.models.MintPermissionResult
 
 data class DialogRequest(
-    val type: DialogRequestType,
+    val group: DialogRequestGroup,
     val results: List<MintPermissionResult>,
     val customContentMapper: DialogContentMapper?,
     val customContentConsumer: DialogContentConsumer?
 )
-
-enum class DialogRequestType {
-    NEEDS_RATIONALE,
-    DENIED,
-    ALL
-}
