+++ flows/src/main/java/ru/mintrocket/lib/mintpermissions/flows/models/DialogRequestGroup.kt
@@ -1,0 +1,6 @@
+package ru.mintrocket.lib.mintpermissions.flows.models
+
+enum class DialogRequestGroup {
+    NEEDS_RATIONALE,
+    DENIED,
+}
